change u64 to u32

+++ backend/model/src/block.rs
@@ -3,13 +3,13 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct EvmBlockInfo {
-    pub number: u64,
+    pub number: u32,
     pub hash: Option<String>,
     pub parent_hash: String,
     pub timestamp: String,
-    pub gas_used: u64,
-    pub gas_limit: u64,
-    pub base_fee_per_gas: Option<u64>,
+    pub gas_used: u32,
+    pub gas_limit: u32,
+    pub base_fee_per_gas: Option<u32>,
     pub validate: String,
     pub extra_data: String,
     pub transactions_count: usize,
