+++ backend/model/src/netwiork.rs
@@ -2,10 +2,10 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct EvmNetworkInfo {
-    pub chain_id: u64,
-    pub gas_price: u64,
-    pub max_priority_fee: Option<u64>,
-    pub max_fee: Option<u64>,
-    pub latest_block_number: u64,
+    pub chain_id: u16,
+    pub gas_price: u32,
+    pub max_priority_fee: Option<u32>,
+    pub max_fee: Option<u32>,
+    pub latest_block_number: u32,
     pub syncing: bool,
 }
